Fix wrong casting

+++ common/src/main/java/us/myles/ViaVersion/protocols/protocol1_13to1_12_2/blockconnections/ConnectionData.java
@@ -185,7 +185,7 @@ public class ConnectionData {
                             block = handler.connect(user, new Position(
                                     (int) (xOff + x),
                                     (short) (yOff + y),
-                                    (short) (zOff + z)
+                                    (int) (zOff + z)
                             ), block);
                             section.setFlatBlock(x, y, z, block);
                         }
